Renamed "internal/service" as "internal/business"
The package service was renamed as business to prevent confusions and
make the architecture style more faithful to the most common way of
doing layered architecture

issue initial-architecture

issue initial-architecture
